BE_API/requirements.txt
Flask
Werkzeug
Flask-SQLAlchemy
marshmallow-sqlalchemy
PyMySQL
python-dotenv
Flask-Marshmallow
marshmallow
SQLAlchemy
google-generativeai
Flask-jwt-extended
